Year 1 ADS&AI repositories/Block D/Group data/requirements.txt
streamlit
streamlit_option_menu
Pillow
pandas
seaborn
matplotlib
plotly
numpy
folium
streamlit_folium
branca
shapely
geopandas
joblib
scikit-learn
